pyproject.toml
[tool.poetry]
name = "mlflow-meetup-lille"
version = "0.1.0"
description = "A demo for the Lille Data Meetup on MLFlow"
authors = ["Kamel Tighidet <kamel.tighidet@cleverconnect.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mlflow = "^2.7.1"
optuna = "^3.4.0"
spacy = "^3.7.2"
nltk = "^3.8.1"
fasttext-wheel = "^0.9.2"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
seaborn = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

